+++ crates/app/src/commands/docker/setup.rs
@@ -31,7 +31,7 @@ pub async fn setup(session: MoonSession) -> SessionResult {
 
     let (action_context, action_graph) = action_graph_builder.build();
 
-    run_action_pipeline(&session, action_context, action_graph).await?;
+    run_action_pipeline(&session, action_context, action_graph, None).await?;
 
     Ok(None)
 }